+++ app/src/main/java/woowacourse/shopping/presentation/cart/CartFragment.kt
@@ -14,11 +14,11 @@ import woowacourse.shopping.R
 import woowacourse.shopping.data.cart.CartRepositoryInjector
 import woowacourse.shopping.data.shopping.ProductRepositoryInjector
 import woowacourse.shopping.databinding.FragmentCartBinding
-import woowacourse.shopping.domain.usecase.DefaultDecreaseCartProductUseCase
-import woowacourse.shopping.domain.usecase.DefaultDeleteCartProductUseCase
-import woowacourse.shopping.domain.usecase.DefaultIncreaseCartProductUseCase
-import woowacourse.shopping.domain.usecase.DefaultLoadCartUseCase
-import woowacourse.shopping.domain.usecase.DefaultLoadPagingCartUseCase
+import woowacourse.shopping.domain.usecase.cart.DefaultDecreaseCartProductUseCase
+import woowacourse.shopping.domain.usecase.cart.DefaultDeleteCartProductUseCase
+import woowacourse.shopping.domain.usecase.cart.DefaultIncreaseCartProductUseCase
+import woowacourse.shopping.domain.usecase.cart.DefaultLoadCartUseCase
+import woowacourse.shopping.domain.usecase.cart.DefaultLoadPagingCartUseCase
 import woowacourse.shopping.presentation.base.BindingFragment
 import woowacourse.shopping.presentation.navigation.ShoppingNavigator
 import woowacourse.shopping.presentation.shopping.ShoppingEventBusViewModel
