+++ app/src/main/java/woowacourse/shopping/presentation/shopping/detail/ProductDetailFragment.kt
@@ -14,7 +14,7 @@ import woowacourse.shopping.R
 import woowacourse.shopping.data.cart.CartRepositoryInjector
 import woowacourse.shopping.data.shopping.ProductRepositoryInjector
 import woowacourse.shopping.databinding.FragmentProductDetailBinding
-import woowacourse.shopping.domain.usecase.DefaultAddCartProductUseCase
+import woowacourse.shopping.domain.usecase.cart.DefaultAddCartProductUseCase
 import woowacourse.shopping.presentation.base.BindingFragment
 import woowacourse.shopping.presentation.navigation.ShoppingNavigator
 import woowacourse.shopping.presentation.shopping.ShoppingEventBusViewModel
