+++ app/src/main/java/woowacourse/shopping/presentation/shopping/product/ProductListFragment.kt
@@ -16,9 +16,9 @@ import woowacourse.shopping.R
 import woowacourse.shopping.data.cart.CartRepositoryInjector
 import woowacourse.shopping.data.shopping.ProductRepositoryInjector
 import woowacourse.shopping.databinding.FragmentProductListBinding
-import woowacourse.shopping.domain.usecase.DefaultDecreaseCartProductUseCase
-import woowacourse.shopping.domain.usecase.DefaultIncreaseCartProductUseCase
-import woowacourse.shopping.domain.usecase.DefaultLoadCartUseCase
+import woowacourse.shopping.domain.usecase.cart.DefaultDecreaseCartProductUseCase
+import woowacourse.shopping.domain.usecase.cart.DefaultIncreaseCartProductUseCase
+import woowacourse.shopping.domain.usecase.cart.DefaultLoadCartUseCase
 import woowacourse.shopping.presentation.base.BindingFragment
 import woowacourse.shopping.presentation.navigation.ShoppingNavigator
 import woowacourse.shopping.presentation.shopping.ShoppingEventBusViewModel
