+++ app/src/test/java/woowacourse/shopping/domain/usecase/DefaultAddCartProductUseCaseTest.kt
@@ -16,6 +16,7 @@ import woowacourse.shopping.domain.entity.Cart
 import woowacourse.shopping.domain.entity.fakeProduct
 import woowacourse.shopping.domain.repository.CartRepository
 import woowacourse.shopping.domain.repository.ProductRepository
+import woowacourse.shopping.domain.usecase.cart.DefaultAddCartProductUseCase
 
 @ExtendWith(MockKExtension::class)
 class DefaultAddCartProductUseCaseTest {
