+++ app/src/test/java/woowacourse/shopping/presentation/cart/CartViewModelTest.kt
@@ -14,11 +14,11 @@ import org.junit.jupiter.api.extension.ExtendWith
 import woowacourse.shopping.domain.entity.Cart
 import woowacourse.shopping.domain.entity.fakeCartProduct
 import woowacourse.shopping.domain.repository.CartRepository
-import woowacourse.shopping.domain.usecase.DecreaseCartProductUseCase
-import woowacourse.shopping.domain.usecase.DeleteCartProductUseCase
-import woowacourse.shopping.domain.usecase.IncreaseCartProductUseCase
-import woowacourse.shopping.domain.usecase.LoadCartUseCase
-import woowacourse.shopping.domain.usecase.LoadPagingCartUseCase
+import woowacourse.shopping.domain.usecase.cart.DecreaseCartProductUseCase
+import woowacourse.shopping.domain.usecase.cart.DeleteCartProductUseCase
+import woowacourse.shopping.domain.usecase.cart.IncreaseCartProductUseCase
+import woowacourse.shopping.domain.usecase.cart.LoadCartUseCase
+import woowacourse.shopping.domain.usecase.cart.LoadPagingCartUseCase
 import woowacourse.shopping.presentation.util.CoroutinesTestExtension
 import woowacourse.shopping.presentation.util.InstantTaskExecutorExtension
 import woowacourse.shopping.presentation.util.getOrAwaitValue
