+++ app/src/test/java/woowacourse/shopping/presentation/shopping/product/ProductListViewModelTest.kt
@@ -13,9 +13,9 @@ import org.junit.jupiter.api.extension.ExtendWith
 import woowacourse.shopping.domain.entity.Cart
 import woowacourse.shopping.domain.entity.fakeProduct
 import woowacourse.shopping.domain.repository.ProductRepository
-import woowacourse.shopping.domain.usecase.DecreaseCartProductUseCase
-import woowacourse.shopping.domain.usecase.IncreaseCartProductUseCase
-import woowacourse.shopping.domain.usecase.LoadCartUseCase
+import woowacourse.shopping.domain.usecase.cart.DecreaseCartProductUseCase
+import woowacourse.shopping.domain.usecase.cart.IncreaseCartProductUseCase
+import woowacourse.shopping.domain.usecase.cart.LoadCartUseCase
 import woowacourse.shopping.presentation.shopping.toShoppingUiModel
 import woowacourse.shopping.presentation.util.CoroutinesTestExtension
 import woowacourse.shopping.presentation.util.InstantTaskExecutorExtension
